Update dependencies after poller directory moving
