+++ pyproject.toml
@@ -69,7 +69,7 @@ filelock = "^3.0.12"
 isort = "^5.10.1"
 
 [tool.poetry.scripts]
-sq-poller = 'suzieq.poller.sq_poller:poller_main'
+sq-poller = 'suzieq.poller.worker.sq_poller:poller_main'
 sq-rest-server = 'suzieq.restServer.sq_rest_server:rest_main'
 sq-coalescer = 'suzieq.utilities.sq_coalescer:coalescer_main'
 sq-anonymizer = 'suzieq.utilities.sq_anonymizer:anonymizer_main'
