+++ suzieq/poller/worker/nodes/__init__.py
@@ -1,3 +1,3 @@
-from suzieq.poller.nodes.node import Node
+from suzieq.poller.worker.nodes.node import Node
 
 __all__ = ['Node']
