+++ suzieq/poller/worker/services/bgp.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 
 from dateparser import parse
 
-from suzieq.poller.services.service import Service
+from suzieq.poller.worker.services.service import Service
 from suzieq.shared.utils import get_timestamp_from_cisco_time
 from suzieq.shared.utils import get_timestamp_from_junos_time
