+++ suzieq/poller/worker/services/devconfig.py
@@ -1,6 +1,6 @@
 import re
 
-from suzieq.poller.services.service import Service
+from suzieq.poller.worker.services.service import Service
 
 
 class ConfigService(Service):
