+++ suzieq/poller/worker/services/inventory.py
@@ -1,6 +1,6 @@
 import re
 import numpy as np
-from suzieq.poller.services.service import Service
+from suzieq.poller.worker.services.service import Service
 
 
 class InventoryService(Service):
