+++ suzieq/poller/worker/services/mlag.py
@@ -1,4 +1,4 @@
-from suzieq.poller.services.service import Service
+from suzieq.poller.worker.services.service import Service
 from suzieq.shared.utils import expand_nxos_ifname
