+++ suzieq/poller/worker/services/ospfNbr.py
@@ -3,7 +3,7 @@ from dateparser import parse
 
 import numpy as np
 
-from suzieq.poller.services.service import Service
+from suzieq.poller.worker.services.service import Service
 from suzieq.shared.utils import get_timestamp_from_cisco_time
 from suzieq.shared.utils import get_timestamp_from_junos_time
